src/app/modules/DPS/requirements.txt
# ── Async / HTTP ──────────────────────────────────────────────────────────────
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
anyio==4.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
sniffio==1.3.1
yarl==1.23.0

# ── Azure ─────────────────────────────────────────────────────────────────────
azure-core==1.38.2
azure-cosmos==4.9.0
azure-eventhub==5.15.1
azure-eventhub-checkpointstoreblob-aio==1.2.0
azure-functions==1.24.0
azure-identity==1.19.0
azure-storage-blob==12.28.0
azurefunctions-extensions-base==1.1.0
azurefunctions-extensions-bindings-cosmosdb==1.0.0b1
isodate==0.7.2
msal==1.35.1
msal-extensions==1.3.1

# ── Elasticsearch ─────────────────────────────────────────────────────────────
elastic-transport==9.2.1
elasticsearch==9.3.0

# ── LangChain / LangGraph ─────────────────────────────────────────────────────
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.11
jsonpatch==1.33
jsonpointer==3.0.0
tenacity==9.1.4

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai==2.24.0
tiktoken==0.12.0
jiter==0.13.0

# ── ML / Embeddings (CPU) ─────────────────────────────────────────────────────
# torch: CPU-only wheel — cuts ~15 GB vs the default CUDA build
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
sentence-transformers==5.2.3
transformers==5.3.0
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.6.0
hf-xet==1.4.0
scikit-learn==1.8.0
scipy==1.17.1
numpy==2.4.3
joblib==1.5.3
threadpoolctl==3.6.0

# ── Streamlit / UI ────────────────────────────────────────────────────────────
streamlit==1.55.0
altair==6.0.0
pydeck==0.9.1
pillow==12.1.1
watchdog==6.0.0
tornado==6.5.5
narwhals==2.18.0
pyarrow==23.0.1
blinker==1.9.0
rich==14.3.3
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
click==8.3.1
typer==0.24.1
shellingham==1.5.4

# ── Pydantic / Config ─────────────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
annotated-types==0.7.0
annotated-doc==0.0.4

# ── Data / Serialisation ──────────────────────────────────────────────────────
pandas==2.3.3
orjson==3.11.7
ormsgpack==1.12.2
PyYAML==6.0.3
toml==0.10.2
protobuf==6.33.5
xxhash==3.6.0
zstandard==0.25.0
fsspec==2026.2.0
filelock==3.25.2
frozenlist==1.8.0
regex==2026.2.28

# ── Crypto / Auth ─────────────────────────────────────────────────────────────
cryptography==46.0.5
cffi==2.0.0
pycparser==3.0
PyJWT==2.11.0
certifi==2026.2.25
charset-normalizer==3.4.4

# ── Misc ──────────────────────────────────────────────────────────────────────
requests==2.32.5
requests-toolbelt==1.0.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3
six==1.17.0
tqdm==4.67.3
distro==1.9.0
packaging==26.0
setuptools==82.0.1
uuid_utils==0.14.1
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
cachetools==7.0.5
mpmath==1.3.0
sympy==1.14.0
networkx==3.6.1
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.3
orjson==3.11.7
urllib3==2.6.3